refactor: HTTP Request 관련 파일들에 대해 패키지 세분화
