+++ src/main/java/util/httprequest/HttpRequestParser.java
@@ -1,4 +1,4 @@
-package util;
+package util.httprequest;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
