+++ src/test/java/util/httprequest/HttpRequestParserTest.java
@@ -1,4 +1,4 @@
-package util;
+package util.httprequest;
 
 import org.junit.jupiter.api.Test;
