+++ src/test/java/util/httprequest/HttpRequestUtilTest.java
@@ -1,4 +1,4 @@
-package util;
+package util.httprequest;
 
 import org.junit.jupiter.api.Test;
